+++ example/lib/src/another_app/home/inherited/controller/inherit_controller.dart
@@ -16,5 +16,6 @@ class InheritController extends StateXController {
       states.first.dependOnInheritedWidget(context);
 
   /// Rebuild the InheritedWidget to also rebuild its dependencies.
+  /// 'first' is the 'root' State object.
   void newAnimals() => states.first.buildInherited();
 }